volume bus control, low pass filter and music transitions
Set up some functions for the volume sliders, just need to call _set_volume() and pass in the float val and the bus type. You can also initialize your sliders with the _get_volume_normalized() function.

In terms of music, you just need to call _change_music and pass in one of the audiostream variables and the transition time in seconds.

With the water toggle you need to call _toggle_water_effect() and pass in a bool.